Revert "refactor(WIP): move common tree logic into a mixin"
This reverts commit 0b2dc1e8ab7f917f5ac90a7e0031faa6fb824571.
