+++ src/ui/layout/RecentObjects.vue
@@ -1,30 +1,23 @@
 <template>
-<div
-    class="c-tree-and-search l-shell__tree"
->
-    <div
-        class="c-tree-and-search__tree c-tree c-tree__scrollable"
-    >
-
-        <tree-item
-            v-for="(recentItem, index) in treeItems"
-            :key="`${recentItem.navigationPath}-recent-${index}`"
-            :node="recentItem"
-            :is-selector-tree="false"
-            :selected-item="selectedItem"
-            :left-offset="recentItem.leftOffset"
-            :is-new="recentItem.isNew"
-            :item-offset="itemOffset"
-            :item-index="index"
-            :item-height="itemHeight"
-            :open-items="openTreeItems"
-            :loading-items="treeItemLoading"
-        />
-        <!-- @tree-item-mounted="scrollToCheck($event)"
-            @tree-item-action="treeItemAction(recentItem, $event)"
-            @tree-item-destroyed="removeCompositionListenerFor($event)"
-            @tree-item-selection="recentItemSelection(recentItem)" -->
-    </div>
+<div class="c-tree-scrollable">
+    <tree-item
+        v-for="(recentItem, index) in recentItems"
+        :key="`${recentItem.navigationPath}-recent-${index}`"
+        :node="recentItem"
+        :is-selector-tree="false"
+        :selected-item="selectedItem"
+        :left-offset="'0px'"
+        :is-new="recentItem.isNew"
+        :item-offset="itemOffset"
+        :item-index="index"
+        :item-height="itemHeight"
+        :open-items="[]"
+        :loading-items="{}"
+    />
+    <!-- @tree-item-mounted="scrollToCheck($event)"
+        @tree-item-destroyed="removeCompositionListenerFor($event)"
+        @tree-item-action="recentItemAction(recentItem, $event)"
+        @tree-item-selection="recentItemSelection(recentItem)" -->
 </div>
 </template>
 
@@ -43,6 +36,10 @@ export default {
     },
     data() {
         return {
+            recentItems: [],
+            selectedItem: null,
+            itemHeight: 27,
+            itemOffset: 0
         };
     },
     async mounted() {
@@ -62,11 +59,11 @@ export default {
         async onHashChange(hash) {
             const objectPath = await this.openmct.objects.getRelativeObjectPath(hash);
             const navigationPath = `/browse/${this.openmct.objects.getRelativePath(objectPath.slice(0, -1))}`;
-            const foundIndex = this.treeItems.findIndex((item) => {
+            const foundIndex = this.recentItems.findIndex((item) => {
                 return navigationPath === item.navigationPath;
             });
             if (foundIndex > -1) {
-                const removedItem = this.treeItems.splice(foundIndex, 1);
+                const removedItem = this.recentItems.splice(foundIndex, 1);
                 this.selectedItem = removedItem[0];
             } else {
                 this.selectedItem = {
@@ -77,18 +74,58 @@ export default {
                 };
             }
 
-            this.treeItems.unshift(this.selectedItem);
-            while (this.treeItems.length > MAX_RECENT_ITEMS) {
-                this.treeItems.pop();
+            this.recentItems.unshift(this.selectedItem);
+            while (this.recentItems.length > MAX_RECENT_ITEMS) {
+                this.recentItems.pop();
             }
         },
-        async loadAndBuildTreeItemsFor(domainObject, parentObjectPath, abortSignal) {
-            let collection = this.openmct.composition.get(domainObject);
-            let composition = await collection.load(abortSignal);
+        calculateHeights() {
+            const RECHECK = 100;
 
-            return composition.map((object) => {
-                return this.buildTreeItem(object, parentObjectPath);
+            return new Promise((resolve, reject) => {
+
+                let checkHeights = () => {
+                    let treeTopMargin = this.getElementStyleValue(this.$refs.mainTree, 'marginTop');
+                    let paddingOffset = 0;
+
+                    if (
+                        this.$el
+                        && this.$refs.search
+                        && this.$refs.mainTree
+                        && this.$refs.treeContainer
+                        && this.$refs.dummyItem
+                        && this.$el.offsetHeight !== 0
+                        && treeTopMargin > 0
+                    ) {
+
+                        this.mainTreeTopMargin = treeTopMargin;
+                        this.mainTreeHeight = this.$el.offsetHeight
+                            - this.$refs.search.offsetHeight
+                            - this.mainTreeTopMargin
+                            - (paddingOffset * 2);
+                        this.itemHeight = this.getElementStyleValue(this.$refs.dummyItem, 'height');
+
+                        resolve();
+                    } else {
+                        setTimeout(checkHeights, RECHECK);
+                    }
+                };
+
+                checkHeights();
             });
+        },
+        getElementStyleValue(el, style) {
+            if (!el) {
+                return;
+            }
+
+            let styleString = window.getComputedStyle(el)[style];
+            let index = styleString.indexOf('px');
+
+            return Number(styleString.slice(0, index));
+        },
+        handleTreeResize() {
+            this.calculateHeights();
         }
     }
 };
